finished limit offset

+++ dropdbase/GpuSqlParser/GpuSqlListener.h
@@ -67,8 +67,8 @@ private:
 public:
 	GpuSqlListener(const std::shared_ptr<Database> &database, GpuSqlDispatcher &dispatcher);
 
-	int32_t resultLimit;
-    int32_t resultOffset;
+	int64_t resultLimit;
+    int64_t resultOffset;
 
     void exitBinaryOperation(GpuSqlParser::BinaryOperationContext *ctx) override;
